% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_utils.R
\name{build_matching_cluster}
\alias{build_matching_cluster}
\title{Build matching clusters from a MatchIt object}
\usage{
build_matching_cluster(mobj)
}
\arguments{
\item{mobj}{A MatchIt object}
}
\value{
integer vector of cluster IDs named by subject ID
}
\description{
Build matching clusters from a MatchIt object
}
\keyword{internal}
