% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{compute_neff}
\alias{compute_neff}
\title{Kish effective sample size}
\usage{
compute_neff(w)
}
\arguments{
\item{w}{Numeric vector of mean-normalised weights (sum = N).}
}
\value{
Scalar Neff.
}
\description{
Kish effective sample size
}
