% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_utils.R
\name{normalize_match_weights}
\alias{normalize_match_weights}
\title{Normalize matching weights}
\usage{
normalize_match_weights(mobj, ids)
}
\arguments{
\item{mobj}{MatchIt object}

\item{ids}{subject IDs in matched data}
}
\value{
normalized numeric vector
}
\description{
Normalize matching weights
}
\keyword{internal}
