% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity_index.R
\name{prepare_heterogeneity}
\alias{prepare_heterogeneity}
\title{Prepare heterogeneity indices}
\usage{
prepare_heterogeneity(omics_type, data = NULL, index = NULL)
}
\arguments{
\item{omics_type}{omics type}

\item{data}{feature matrix (samples x features)}

\item{index}{optional user-supplied heterogeneity matrix}
}
\value{
data.frame of heterogeneity covariates
}
\description{
Prepare heterogeneity indices
}
