% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{smd_summary}
\alias{smd_summary}
\title{Summarise pre/post matching SMD}
\usage{
smd_summary(pre_smd, post_smd)
}
\arguments{
\item{pre_smd}{Named numeric vector of pre-matching SMD values.}

\item{post_smd}{Named numeric vector of post-matching SMD values.}
}
\value{
Data frame with columns \code{covariate}, \code{smd_pre},
\code{smd_post}.
}
\description{
Summarise pre/post matching SMD
}
