% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{toy_heterogeneity}
\alias{toy_heterogeneity}
\title{Simulated heterogeneity indices}
\format{
A data frame with 200 rows and 2 columns:
\describe{
\item{het1}{First principal component score.}
\item{het2}{Second principal component score.}
}
}
\usage{
toy_heterogeneity
}
\description{
Data frame with 200 rows and two columns (\code{het1}, \code{het2}) derived
from PCA of \code{\link{toy_omics}}. Used as propensity score covariates in
\code{\link{MatchOmics}}.
}
\examples{
data(toy_heterogeneity)
head(toy_heterogeneity)
}
\keyword{datasets}
