% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{toy_marker}
\alias{toy_marker}
\title{Toy marker vector for MatchOmics examples}
\format{
A numeric vector of length 120.
}
\usage{
toy_marker
}
\description{
Simulated continuous marker used in package examples.
}
\keyword{datasets}
