% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{toy_omics}
\alias{toy_omics}
\title{Simulated proteomics matrix}
\format{
A numeric matrix with 50 rows and 200 columns.
}
\usage{
toy_omics
}
\description{
A 50 × 200 numeric matrix of log-normalised protein abundances for 200
simulated subjects and 50 proteins. Rows are proteins, columns are subjects.
}
\examples{
data(toy_omics)
dim(toy_omics)
}
\keyword{datasets}
