% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{toy_outcome}
\alias{toy_outcome}
\title{Simulated binary outcome}
\format{
An integer vector of length 200.
}
\usage{
toy_outcome
}
\description{
Integer vector of length 200 (0/1) representing disease status for the
subjects in \code{\link{toy_omics}}.
}
\examples{
data(toy_outcome)
table(toy_outcome)
}
\keyword{datasets}
