% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{toy_omics}
\alias{toy_omics}
\title{Toy omics matrix for MatchOmics examples}
\format{
A numeric matrix with 120 rows and 20 columns.
}
\usage{
toy_omics
}
\description{
Simulated omics feature matrix.
}
\keyword{datasets}
