% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{toy_outcome}
\alias{toy_outcome}
\title{Toy binary outcome for MatchOmics examples}
\format{
An integer vector of length 120 with values 0 and 1.
}
\usage{
toy_outcome
}
\description{
Simulated binary outcome corresponding to \code{toy_marker}.
}
\keyword{datasets}
